papp-ai-docs/pyproject.toml
[project]
name = "papp-ai-docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "adal>=1.2.7",
    "azure-functions>=1.24.0",
    "diff-match-patch>=20241021",
    "dynamics365crm-python>=1.0.2",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "mammoth>=1.11.0",
    "pandas>=2.3.3",
    "pydocx>=0.9.10",
    "python-docx>=1.2.0",
]
